remove msbuild-based enable package restore command code
