+++ src/NuGet.Core/NuGet.PackageManagement/IDE/IPackageRestoreManager.cs
@@ -5,29 +5,12 @@ using System;
 using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
-using NuGet.Packaging.Core;
 using NuGet.ProjectManagement;
 
 namespace NuGet.PackageManagement
 {
     public interface IPackageRestoreManager
     {
-        /// <summary>
-        /// Gets a value indicating whether the current solution is configured for Package Restore mode.
-        /// </summary>
-        [Obsolete("Enabling and querying legacy package restore is not supported in VS 2015 RTM.")]
-        bool IsCurrentSolutionEnabledForRestore { get; }
-
-        /// <summary>
-        /// Configures the current solution for Package Restore mode.
-        /// </summary>
-        /// <param name="fromActivation">
-        /// if set to <c>false</c>, the method will not show any error message, and will
-        /// not set package restore consent.
-        /// </param>
-        [Obsolete("Enabling and querying legacy package restore is not supported in VS 2015 RTM.")]
-        void EnableCurrentSolutionForRestore(bool fromActivation);
-
         /// <summary>
         /// Occurs when it is detected that the packages are missing or restored for the current solution.
         /// </summary>
@@ -98,36 +81,5 @@ namespace NuGet.PackageManagement
             IEnumerable<PackageRestoreData> packages,
             INuGetProjectContext nuGetProjectContext,
             CancellationToken token);
-    }
-
-    /// <summary>
-    /// If 'Restored' is false, it means that the package was already restored
-    /// If 'Restored' is true, the package was restored and successfully
-    /// </summary>
-    public class PackageRestoredEventArgs : EventArgs
-    {
-        public PackageIdentity Package { get; private set; }
-        public bool Restored { get; private set; }
-
-        public PackageRestoredEventArgs(PackageIdentity packageIdentity, bool restored)
-        {
-            if (packageIdentity == null)
-            {
-                throw new ArgumentNullException("packageIdentity");
-            }
-
-            Package = packageIdentity;
-            Restored = restored;
-        }
-    }
-
-    public class PackagesMissingStatusEventArgs : EventArgs
-    {
-        public bool PackagesMissing { get; private set; }
-
-        public PackagesMissingStatusEventArgs(bool packagesMissing)
-        {
-            PackagesMissing = packagesMissing;
-        }
     }
 }
