+++ src/NuGet.Core/NuGet.PackageManagement/IDE/PackagesMissingStatusEventArgs.cs
@@ -1,0 +1,17 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+
+namespace NuGet.PackageManagement
+{
+    public class PackagesMissingStatusEventArgs : EventArgs
+    {
+        public bool PackagesMissing { get; private set; }
+
+        public PackagesMissingStatusEventArgs(bool packagesMissing)
+        {
+            PackagesMissing = packagesMissing;
+        }
+    }
+}
